+++ lib/types/scene.ts
@@ -4,7 +4,7 @@ export type Color3 = [number, number, number];
 
 export type RenderMode = "texture" | "color" | "none";
 
-export type AnimationProp = "pos" | "rot" | "scale" | "visible";
+export type AnimationProp = "pos" | "rot" | "scale" | "visible" | "tex";
 
 export type Axis = "x" | "y" | "z";
 
@@ -17,6 +17,7 @@ export interface Transform {
 export interface Face {
 	vertexIndices: number[];
 	uvs: Float32Array;
+	staticUvs: Float32Array;
 	color: number;
 	doubleSided: boolean;
 	priority: boolean;
@@ -38,6 +39,10 @@ export interface AnimationClip {
 	times?: number | undefined;
 	curve: string;
 	pingpong: boolean;
+	faceIndex?: number | undefined;
+	frames?: number | undefined;
+	step?: number | undefined;
+	returnUv?: boolean | undefined;
 }
 
 export interface MotionData {
@@ -48,14 +53,18 @@ export interface SceneNode {
 	name: string;
 	visible: boolean;
 	renderVisible: boolean;
+	ghost: boolean;
 	children: SceneNode[];
 	transform: Transform;
 	staticTransform: Transform;
 	originalVisible: boolean;
 	mesh: Mesh | null;
 	motions: MotionData;
+	hasTexClips: boolean;
+	uvsDirty: boolean;
 	dirty: boolean;
 	localMatrix: mat4;
+	worldMatrix: mat4;
 }
 
 export interface TextureData {
@@ -77,6 +86,7 @@ export interface ExportSettings {
 	cameraModeDirection: "left" | "right";
 	cameraModeSpeed: number;
 	animate: boolean;
+	animateLoops: number;
 	outlineSize: number;
 	outlineColor: Color3;
 	scanlines: boolean;